scripts/news-monitor/requirements.txt
anthropic>=0.25.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
